requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
build==1.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.15.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
pip-tools==7.4.1
proglog==0.1.10
protobuf==4.25.3
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pymongo==4.7.3
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
stripe==9.12.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf_keras==2.16.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
